保存预测结果——后端

+++ bigData/src/main/java/com/xiaoke_1256/orders/bigdata/product/dto/ProductPredictInput.java
@@ -13,7 +13,7 @@ public class ProductPredictInput {
     /**
      * 模型主键
      */
-    private Integer modelId;
+    private Long modelId;
 
     /**
      * 系数
@@ -37,11 +37,11 @@ public class ProductPredictInput {
         this.modelPath = modelPath;
     }
 
-    public Integer getModelId() {
+    public Long getModelId() {
         return modelId;
     }
 
-    public void setModelId(Integer modelId) {
+    public void setModelId(Long modelId) {
         this.modelId = modelId;
     }
